refactoring apps to dedicated project
