+++ src/modules/cmdpal/Exts/Microsoft.CmdPal.Ext.Apps/AppCache.cs
@@ -2,9 +2,10 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using AllApps.Programs;
+using System;
+using System.Collections.Generic;
 
-namespace WindowsCommandPalette.BuiltinCommands.AllApps;
+namespace Microsoft.CmdPal.Ext.Apps.Programs;
 
 public sealed class AppCache
 {
