+++ src/modules/cmdpal/Exts/Microsoft.CmdPal.Ext.Apps/Programs/AppxFactory.cs
@@ -4,7 +4,7 @@
 
 using System.Runtime.InteropServices;
 
-namespace AllApps.Programs;
+namespace Microsoft.CmdPal.Ext.Apps.Programs;
 
 // Reference : https://stackoverflow.com/questions/32122679/getting-icon-of-modern-windows-app-from-a-desktop-application
 [Guid("5842a140-ff9f-4166-8f5c-62f5b7b0c781")]
