+++ src/modules/cmdpal/Exts/Microsoft.CmdPal.Ext.Apps/Programs/IPackageManager.cs
@@ -2,7 +2,9 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace AllApps.Programs;
+using System.Collections.Generic;
+
+namespace Microsoft.CmdPal.Ext.Apps.Programs;
 
 public interface IPackageManager
 {
