+++ src/modules/cmdpal/Exts/Microsoft.CmdPal.Ext.Apps/Programs/IShellLinkHelper.cs
@@ -2,7 +2,7 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace WindowsCommandPalette.BuiltinCommands.AllApps;
+namespace Microsoft.CmdPal.Ext.Apps.Programs;
 
 public interface IShellLinkHelper
 {
