+++ src/modules/cmdpal/Exts/Microsoft.CmdPal.Ext.Apps/Programs/Native.cs
@@ -2,10 +2,11 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Runtime.InteropServices;
 using System.Text;
 
-namespace AllApps.Programs;
+namespace Microsoft.CmdPal.Ext.Apps.Programs;
 
 internal sealed class Native
 {
