+++ src/modules/cmdpal/Exts/Microsoft.CmdPal.Ext.Apps/Programs/ShellLinkHelper.cs
@@ -2,11 +2,12 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.ComTypes;
 using System.Text;
 
-namespace WindowsCommandPalette.BuiltinCommands.AllApps;
+namespace Microsoft.CmdPal.Ext.Apps.Programs;
 
 public class ShellLinkHelper : IShellLinkHelper
 {
