+++ src/modules/cmdpal/WindowsCommandPalette.sln
@@ -48,6 +48,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.CmdPal.UI.ViewMod
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UI", "UI", "{865E9369-C53A-40B8-829C-F7843F66D3A1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.CmdPal.Ext.Apps", "Exts\Microsoft.CmdPal.Ext.Apps\Microsoft.CmdPal.Ext.Apps.csproj", "{2012F5FE-BF53-4E51-868D-0E5A74C11878}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|ARM64 = Debug|ARM64
@@ -318,6 +320,18 @@ Global
 		{D2FC419D-0ABC-425F-9D43-A7782AC4A0AE}.Release|x64.Build.0 = Release|x64
 		{D2FC419D-0ABC-425F-9D43-A7782AC4A0AE}.Release|x86.ActiveCfg = Release|x64
 		{D2FC419D-0ABC-425F-9D43-A7782AC4A0AE}.Release|x86.Build.0 = Release|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Debug|ARM64.Build.0 = Debug|ARM64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Debug|x64.ActiveCfg = Debug|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Debug|x64.Build.0 = Debug|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Debug|x86.ActiveCfg = Debug|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Debug|x86.Build.0 = Debug|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Release|ARM64.ActiveCfg = Release|ARM64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Release|ARM64.Build.0 = Release|ARM64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Release|x64.ActiveCfg = Release|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Release|x64.Build.0 = Release|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Release|x86.ActiveCfg = Release|x64
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878}.Release|x86.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -341,6 +355,7 @@ Global
 		{399B53F1-5AA6-4FE0-8C3C-66E07B84E6F1} = {B7FF739F-7716-4FC3-B622-705486187B87}
 		{1DF70F56-ABB2-4798-BBA5-0B9568715BA1} = {865E9369-C53A-40B8-829C-F7843F66D3A1}
 		{D2FC419D-0ABC-425F-9D43-A7782AC4A0AE} = {865E9369-C53A-40B8-829C-F7843F66D3A1}
+		{2012F5FE-BF53-4E51-868D-0E5A74C11878} = {272D0E9A-8FC3-49F5-8FAD-79ABAE8AB1E4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BC94BFC2-A741-4978-B6A4-9E01B7660E6B}
