+++ src/modules/cmdpal/WindowsCommandPalette/Views/MainViewModel.xaml.cs
@@ -5,6 +5,7 @@
 using System.Collections.ObjectModel;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.WindowsRuntime;
+using Microsoft.CmdPal.Ext.Apps.Programs;
 using Microsoft.CmdPal.Ext.Bookmarks;
 using Microsoft.CmdPal.Ext.Calc;
 using Microsoft.CmdPal.Ext.Settings;
@@ -12,7 +13,6 @@ using Microsoft.CmdPal.Extensions;
 using Microsoft.CmdPal.Extensions.Helpers;
 using Windows.Foundation;
 using WindowsCommandPalette.BuiltinCommands;
-using WindowsCommandPalette.BuiltinCommands.AllApps;
 using WindowsCommandPalette.Models;
 
 namespace WindowsCommandPalette.Views;
